+++ CHANGELOG.md
@@ -10,8 +10,10 @@ All notable changes to this project will be documented in this file.
 
 ### Fixes
 
+- [Typescript] Added type definition for `setGenerator`.
 - [Typescript] Fixed types of decorators.
 - [Typescript] Add Tests to test Types.
+- Fixed typo in learn-to-use docs.
 
 ## 0.16 - 2020-03-06